pyproject.toml
[project]
name = "ai-side"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "fastapi>=0.128.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
